game: Define the Cell type
A cell is the smallest unit in the Life2 simulation game.  It carries
information about its aliveness (whether or not it is filled), and the
team it is part of (TEAM_A or TEAM_B).

The cell type is currently implemented as an enumeration of integers,
merging the two states in one value.  I believe integers are the fastest
type JavaScript runtimes can manipulate, especially at this small scale
(unsigned 8-bit integers).  Given these properties, I believe using
integers will avoid huge performance and memory consumption in the
browser when dealing with boards of larger sizes (or even pseudo
infinite sizes...).

Signed-off-by: Nicolas Paul <n@nc0.fr>