game: Add Board abstraction
The Board class is an abstraction over any 2D shaped board to allow the
use of any two dimensional shapes as a grid for the Life2 simulation.

This patch creates and exposes two types: the Board class, and a Grid
typedef.

The Board class is expected to be inherited by more concrete shaped
boards, such as classic "Rectangular" boards, or more complex "Triangle"
ones.  Boards act as their own abstraction over their shape, with the
only requirement of implementing the small Board API defined by the
inherited class.

The Grid type serves as a encoded representation of the Board shape
using JavaScript Arrays and our defined Life2.Cell enumeration:

    import {Cell} from "@life2/game";

    /**
     * @typedef {!Array<!Array<!Cell>>}
     */
    let Grid;

The Grid structure can be though as a protocol.

The two combined can be represented as the following:

      BOARD                                  CODE

    ##########      const grid = [[3, 3, 3, 3, 3, 3, 3, 3, 3, 3],
    ###a.b.###                    [3, 3, 3, 1, 0, 2, 0, 3, 3, 3],
    #.aab..ba#  ->                [3, 0, 1, 1, 2, 0, 0, 2, 1, 3],
    ###bb..###                    [3, 3, 3, 2, 2, 0, 0, 3, 3, 3],
    ##########                    [3, 3, 3, 3, 3, 3, 3, 3, 3, 3]];

    Assuming:

      * . = 0 = Cell.EMPTY
      * a = 1 = Cell.TEAM_A
      * b = 2 = Cell.TEAM_B
      * # = 3 = Cell.BARRIER

The Grid representation could be obtained by calling the getGrid()
method, which can then be used by some user interface:

  1.  The simulator interface (our @life2/simulator)
  2.  A text format to share the grid state offline (our @life2/l2sf).

Signed-off-by: Nicolas Paul <n@nc0.fr>