Remove the state computation logic from Board
This patch removes the logic of computing the next state of a board to
the Board class.  Instead, this will be the responsability of the World
class.

The rational behind this change is that it will simplifies the reasoning
about the Board as only being a geometrical data structure, leaving
the class to a single responsability.  On the other hand, the World
class, which is already the manager of the game loop, will see its logic
simplified by being able to directly compute the next state, instead of
relying on its dependency on a Board.

Signed-off-by: Nicolas Paul <n@nc0.fr>