Remove cyclic dependecy import
The rule.js file was imported the Cell type from index.js instead of
cell.js, creating a cyclic import graph as index.js itself imports the
rule.js file.  While not illegal in the Node.js ecosystem, it is still
an issue that could crash some static analysis tools, and a bad pattern
to avoid in general.

Signed-off-by: Nicolas Paul <n@nc0.fr>