Add RulesManager helper
RulesManager is a utility class that has the responsability of storing
and managing a set of rules.  Rules are identified by their name,
meaning that there cannot be two rules with the same names.  However, no
check is done on the function content:

    const func = (cell, _) => cell;
    const rule1 = new Rule("foo", "", func);
    const rule2 = new Rule("bar", "", func);
    const rule3 = new Rule("foo", "", func);

    const manager = new RulesManager();
    manager.add(rule1);  // OK
    manager.add(rule2);  // OK because "foo" != "bar"
    manager.add(rule3);  // Error("Rule 'foo' is already registered')

Also, the manager internally use a JavaScript Map object, meaning the
order returned when getting all registered rules is not promised to be
the same has the order of registration.  This also means that the
performance and time complexity of the object is dependant on the
JavaScript engine (the ECMA Script specification only imposes a
datastructure where operations have a complexity better than O(n)).
I believe it is a non-issue as the manager should not keep a lot of
rules anyway.

The reason for the existance of the RulesManager class is to simplify
the facade that is the World class.

Signed-off-by: Nicolas Paul <n@nc0.fr>